day and week view fixed. logout on navbar
