+++ familytabs/src/Components/Navigation.js
@@ -2,6 +2,22 @@ import React from 'react';
 import {Link} from 'react-router-dom'
 import styled from "styled-components";
 
+const Buttons = styled.button`
+    border: 4px solid orange;
+    background-color: skyblue;
+    font-family: 'Lato', sans-serif;
+    display: flex;
+    align-items: center;
+    color: white;
+    font-weight: 300;
+    border: 1px solid #86AEB1;
+    width: 100%;
+    height: 30px;
+    padding-left: 15px;
+    border-radius: 0.1rem;
+    text-decoration: none;
+    font-size: 12px;
+`;
 
 const NavBarStyling = styled.div`
     border: 2px solid orange;
@@ -61,7 +77,8 @@ class Navigation extends React.Component {
             <Links to='/home/household'>Household</Links>
             <Links to='/home/billing' className="nav_button">Billing</Links>
         </NavButtonsContainer>
-       </NavBarStyling>
+       <Buttons onClick={this.props.auth.logout}>Log out</Buttons>
+        </NavBarStyling>
     )
   }
   }
